produce-tweets/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
kafka-python==1.4.7
requests==2.22.0
urllib3==1.25.7
